Move keys to constants
